test: split dashboard payload tests
